fix: restore GUC struct definitions for find_option and add defining_search_path to catalog
The pg_sys config_generic struct doesn't provide the nested access pattern needed
for GUC stack traversal. Restore custom struct definitions for GucVariable, GucStack,
and GucValue to properly access the search_path from the GUC stack.

Also add defining_search_path column to the expected catalog schema in upgrade tests,
which is part of the LSEC-1 security context foundation (v0.87.7).

+++ src/api/security_context.rs
@@ -150,7 +150,7 @@ fn saved_pre_definer_search_path() -> Result<String, PgTrickleError> {
             ));
         }
         let guc_var = guc_var as *mut GucVariable;
-        if (*guc_var).vartype != pg_sys::config_type::PGC_STRING {
+        if (*guc_var).vartype != pg_sys::config_type::PGC_STRING as u32 {
             return Err(PgTrickleError::InternalError(
                 "security context: search_path GUC has an unexpected type".into(),
             ));
